requirements.txt
Keras==2.2.5
opencv-contrib-python==3.4.2.17
opencv-python~=4.2.0.32
Pillow~=5.1.0
tensorflow-gpu~=2.1.0
progressbar2
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
tqdm==4.28.1
numpy~=1.18.1
Cython~=0.29.15
six~=1.14.0